Product CRUD: Delete is now done

+++ src/main/java/org/auctionproject/web/repository/ProductRepository.java
@@ -3,17 +3,14 @@ package org.auctionproject.web.repository;
 import org.auctionproject.web.model.Product;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
-import org.springframework.data.repository.CrudRepository;
 import org.springframework.data.repository.PagingAndSortingRepository;
 import org.springframework.stereotype.Repository;
-
-import java.util.List;
 
 /**
  * Created by bishal on 4/3/17.
  */
 @Repository
-public interface ProductRepository extends PagingAndSortingRepository<Product, Long> {
+public interface ProductRepository extends PagingAndSortingRepository<Product, Integer> {
 
     Page<Product> findAll(Pageable pageRequest);
 }
